1061: simplify CLI test build orchestration
