+++ test/cli_acceptance_matrix_strictness.test.ts
@@ -3,7 +3,8 @@ import { mkdir, mkdtemp, readFile, rm, writeFile } from 'node:fs/promises';
 import { tmpdir } from 'node:os';
 import { join, resolve } from 'node:path';
 
-import { ensureCliBuilt, exists, normalizePathForCompare, runCli } from './helpers/cli.js';
+import { ensureCliBuilt } from './helpers/cliBuild.js';
+import { exists, normalizePathForCompare, runCli } from './helpers/cli.js';
 
 type ArtifactKind = 'bin' | 'hex' | 'd8m' | 'lst';
 type ArtifactExpectation = Record<ArtifactKind, boolean>;
