+++ test/cli_artifacts.test.ts
@@ -7,7 +7,8 @@ import { fileURLToPath } from 'node:url';
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = dirname(__filename);
 
-import { ensureCliBuilt, exists, runCli } from './helpers/cli.js';
+import { ensureCliBuilt } from './helpers/cliBuild.js';
+import { exists, runCli } from './helpers/cli.js';
 
 describe('cli artifacts', () => {
   beforeAll(async () => {
