+++ test/cli_zax_smoke.test.ts
@@ -1,8 +1,13 @@
-import { describe, expect, it } from 'vitest';
+import { beforeAll, describe, expect, it } from 'vitest';
 
+import { ensureCliBuilt } from './helpers/cliBuild.js';
 import { runCli } from './helpers/cli.js';
 
 describe('npm zax smoke', () => {
+  beforeAll(async () => {
+    await ensureCliBuilt();
+  });
+
   it('builds and prints version via npm script wrapper', async () => {
     const res = await runCli(['--version']);
     expect(res.code).toBe(0);
